changing c4d variable read function to be more pythonic.
Notes per Maxon support:
These lines does not perform the same operation.
doc.GetDataInstance() returns the data container instance for the document. While doc[plugin_ID] returns the plugin_ID sub-container of the document data container.
With the data container instance another line is needed to retrieve the sub-container.
Using the [] operator in Python to access parameters is preferred because it is faster to write and more pythonic. Unfortunately it is broken in R18.057.